--- ofsencryptors/internal/messages/messages.go
@@ -1,4 +1,0 @@
-package messages
-
-const ERR_AES_CIPHERTEXT_NONCE_SIZE string = "ciphertext length less than nonce size"
-const ERR_AES_KEYLEN string = "aes key length must be 16, 24 or 32 bytes long"
